+++ utils/slots.go
@@ -2,8 +2,8 @@ package utils
 
 import "time"
 
-func CaltulateFirstSlot(timestamp time.Time, slotInverval uint8, activeDevegates int) int64 {
+func CaltulateFirstSlot(timestamp time.Time, slotInverval uint8, activeDevegates uint16) uint64 {
 	slot := (timestamp.Unix() - int64(slotInverval)) / int64(slotInverval)
 	diff := slot % int64(activeDevegates)
-	return slot - diff*int64(slotInverval)
+	return uint64(slot - diff*int64(slotInverval))
 }
